overhaul metric.jl, reduce useful_functions.jl
