+++ src/RelativisticDynamics.jl
@@ -16,7 +16,7 @@ include("system_parameters.jl")
 include("useful_functions.jl")
 include("metric.jl")
 include("universal_constants.jl")
-include("model.jl")
+#include("model.jl")
 include("initial_conditions.jl")
 include("timestepping.jl")
 include("orbit.jl")
