+++ src/universal_constants.jl
@@ -88,93 +88,5 @@ end
 
 
 
-"""
-    E,L,Q = ELQ(a,α,e,ι,D)
-Calculate the energy, angular momentum and Carter constant given the Keplerian orbital parameters, and the BH spin/direction
-"""
-function ELQ(a,α,e,ι,D)
-    
-
-    zm = cos(ι)
-
-    r_periapsis= α*(1-e)
-    r_apoapsis = α*(1+e)
-    
-    #Define some orbital coefficients
-    f1 = mapping_f(r_periapsis,a,zm)
-    g1 = mapping_g(r_periapsis,a)
-    h1 = mapping_h(r_periapsis,a,zm)
-    d1 = mapping_d(r_periapsis,a,zm)
-
-    f2 = mapping_f(r_apoapsis,a,zm)
-    g2 = mapping_g(r_apoapsis,a)
-    h2 = mapping_h(r_apoapsis,a,zm)
-    d2 = mapping_d(r_apoapsis,a,zm)
-
- 
-    #Determinants 
-    κ = d1*h2 - d2*h1 
-    ϵ = d1*g2 - d2*g1 
-    ρ = f1*h2 - f2*h1
-    η = f1*g2 - f2*g1 
-    σ = g1*h2 - g2*h1 
-
-
-    #Energy
-    E_numerator   = κ*ρ+2.0*η*σ-2.0*D*sqrt(σ*(σ*ϵ^2 + ρ*ϵ*κ-η*κ^2))
-    E_denominator = ρ^2 + 4.0*η*σ
-    E = sqrt(E_numerator/E_denominator)
-
-    #Angular momentum 
-    L = -g1*E/h1 + D*sqrt(g1^2 * E^2 +(f1*E^2 -d1)*h1)/h1
-
-    #Carter 
-    Q = zm^2*(a^2*(1.0-E^2)+L^2/(1.0-zm^2))
-
-    return E,L,Q
-
-
-end 
-
-
-
-"""
-    f = mapping_f(r,a,zminus)
-Mapping function `f` used when converting from Keplerian orbital parameters to constants of motion
-"""
-function mapping_f(r,a,zminus)
-Δ = delta(r,a)
-return r^4 +a^2 * (r*(r+2.0) + zminus^2 * Δ)
-end
-
-
-"""
-    g = mapping_g(r,a)
-Mapping function `g` used when converting from Keplerian orbital parameters to constants of motion
-"""
-function mapping_g(r,a)
-return 2*a*r
-end
-
-"""
-    h = mapping_h(r,a,zminus)
-Mapping function `h` used when converting from Keplerian orbital parameters to constants of motion
-"""
-function mapping_h(r,a,zminus)
-Δ = delta(r,a)
-return r*(r-2.0) + (zminus^2 * Δ)/(1.0 - zminus^2)
-end
-
-
-"""
-    d = mapping_d(r,a,zminus)
-Mapping function `d` used when converting from Keplerian orbital parameters to constants of motion
-"""
-function mapping_d(r,a,zminus)
-Δ = delta(r,a)
-return (r^2 +a^2 * zminus^2)*Δ
-end
-
-
