+++ src/useful_functions.jl
@@ -1,20 +1,6 @@
 using LinearAlgebra
 using ChainRulesCore
-"""
-    Δ = delta(r,a)
-The well-known delta function of the Kerr metric
-"""
-function delta(r,a)
-return r^2 -2.0*r + a^2
-end 
 
-"""
-    Σ = sigma(r,θ,a)
-The well-known sigma function of the Kerr metric
-"""
-function sigma(r,θ,a)
-return r^2 + a^2 * cos(θ)^2
-end 
 
 """
     p_{μ} = convert_to_covariant(metric,p^{μ})
@@ -29,25 +15,11 @@ function convert_to_covariant(metric,vector)
 end 
 
 
-
 """
-    K =  Kretschmann_scalar(r,θ,a)
-Kretschman scalar for the Kerr metric
-"""
-function Kretschmann_scalar(r,θ,a)
-
-    Σ = sigma(r,θ,a)
-
-    return 48.0*(2.0*r^2-Σ)*(Σ^2-16.0*r^2*a^2*cos(θ)^2)/Σ^6
-
-end 
-
-
-"""
-    l = calculate_levi(NF)
+    l = levi_civita_symbol(NF)
 Determine the Levi-civita psuedo tensor
 """
-function calculate_levi()
+function levi_civita_symbol()
 
     # Levi civita tensor
     levi = zeros(Integer,4,4,4,4) 
@@ -69,17 +41,13 @@ function calculate_levi()
     return levi 
 end 
 
-
-
-
-
 """
-    ϵ = permutation_tensor(metric)
+    ϵ = levi_civita_tensor(metric)
 Calcualte the Levi-civita tensor in an arbitrary basis.
 """
-function permutation_tensor(metric)
+function levi_civita_tensor(metric)
 
-    ϵ = calculate_levi()
+    ϵ = levi_civita_symbol()
 
     det_g = det(metric)
